add Haptic
추후 sin 함수 이용해서 조금 더 찰진맛이 나도록 구현해야함.
